Rugguard bot/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rugguard-bot"
version = "1.0.0"
description = "X bot for analyzing account trustworthiness in the Solana ecosystem"
authors = [{name = "Project RUGGUARD"}]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "tweepy==4.14.0",
    "python-dotenv==1.0.0", 
    "requests==2.31.0",
    "textblob==0.17.1"
]

[project.scripts]
rugguard-bot = "main:main"
